Move buffer package to internal/buffer
